acu: Move ACU agent into library and setup plugin file
