+++ agents/ocs_plugin_so.py
@@ -17,7 +17,6 @@ for n, f in [
         ('BlueforsAgent', 'bluefors/bluefors_log_tracker.py'),
         ('CryomechCPAAgent', 'cryomech_cpa/cryomech_cpa_agent.py'),
         ('LATRtXYStageAgent', 'xy_stage/xy_latrt_agent.py'),
-        ('ACUAgent', 'acu/acu_agent.py'),
         ('FTSAerotechAgent', 'fts_aerotech_stage/fts_aerotech_agent.py'),
         ('VantagePro2Agent', 'vantagePro2_agent/vantage_pro2_agent.py'),
         ('HWPPicoscopeAgent', 'hwp_picoscope/pico_agent.py'),
