+++ docs/agents/acu_agent.rst
@@ -11,9 +11,9 @@ It is used for readout of encoder measurements and control of telescope
 platforms.
 
 .. argparse::
-    :filename: ../agents/acu/acu_agent.py
+    :filename: ../socs/agents/acu/agent.py
     :func: add_agent_args
-    :prog: python3 acu_agent.py
+    :prog: python3 agent.py
 
 Dependencies
 ------------
@@ -85,7 +85,7 @@ example configuration block is below::
 Agent API
 ---------
 
-.. autoclass:: agents.acu.acu_agent.ACUAgent
+.. autoclass:: socs.agents.acu.agent.ACUAgent
     :members:
 
 Example Clients
@@ -109,5 +109,5 @@ acquisition processes are running::
 Supporting APIs
 ---------------
 
-.. automodule:: agents.acu.scan_helpers
+.. automodule:: socs.agents.acu.drivers
     :members:
