+++ socs/plugin.py
@@ -1,0 +1,4 @@
+package_name = 'socs'
+agents = {
+    'ACUAgent': {'module': 'socs.agents.acu.agent', 'entry_point': 'main'},
+}
